refactor: file names to kebab
