+++ src/features/admin/evaluation-administrations/[id]/evaluees/[evaluation_result_id]/evaluators/evaluators-roles.tsx
@@ -1,9 +1,9 @@
 import { useNavigate, useParams } from "react-router-dom"
 import { useEffect } from "react"
-import { getEvaluationTemplates } from "../../../../../../../redux/slices/evaluationTemplatesSlice"
+import { getEvaluationTemplates } from "../../../../../../../redux/slices/evaluation-templates-slice"
 import { useAppSelector } from "../../../../../../../hooks/useAppSelector"
 import { useAppDispatch } from "../../../../../../../hooks/useAppDispatch"
-import { getEvaluationResult } from "../../../../../../../redux/slices/evaluationResultSlice"
+import { getEvaluationResult } from "../../../../../../../redux/slices/evaluation-result-slice"
 import { Menu } from "../../../../../../../components/shared/Menu"
 
 export const EvaluatorsRoles = () => {
