+++ src/pages/admin/evaluation-administrations/create/create_evaluation.tsx
@@ -1,6 +1,6 @@
 import { useTitle } from "../../../../hooks/useTitle"
 import { CreateEvaluationHeader } from "../../../../features/admin/evaluation-administrations/create/CreateEvaluationHeader"
-import { CreateEvaluationForm } from "../../../../features/admin/evaluation-administrations/create/CreateEvaluationForm"
+import { CreateEvaluationForm } from "../../../../features/admin/evaluation-administrations/create/create-evaluation-form"
 
 export default function CreateEvaluation() {
   useTitle("Create Evaluation")
