+++ src/pages/auth/login.tsx
@@ -1,6 +1,6 @@
 import { GoogleOAuthProvider } from "@react-oauth/google"
-import { LoginForm } from "../../features/auth/login/LoginForm"
-import { LoginOptions } from "../../features/auth/login/LoginOptions"
+import { LoginForm } from "../../features/auth/login/login-form"
+import { LoginOptions } from "../../features/auth/login/login-options"
 import { useTitle } from "../../hooks/useTitle"
 
 export default function Login() {
