+++ src/redux/slices/evaluation-templates-slice.ts
@@ -5,7 +5,7 @@ import {
   type EvaluationTemplateFilters,
   type EvaluationTemplate,
 } from "../../types/evaluation-template-type"
-import { axiosInstance } from "../../utils/axiosInstance"
+import { axiosInstance } from "../../utils/axios-instance"
 import { Loading } from "../../types/loadingType"
 
 export const getEvaluationTemplates = createAsyncThunk(
