+++ src/redux/slices/external-users-slice.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit"
 import { type AxiosError } from "axios"
 import { type ApiError } from "../../types/apiErrorType"
-import { axiosInstance } from "../../utils/axiosInstance"
+import { axiosInstance } from "../../utils/axios-instance"
 import { Loading } from "../../types/loadingType"
 import { type ExternalUser, type ExternalUserFilters } from "../../types/external-user-type"
 import { type ExternalUserFormData } from "../../types/formDataType"
