+++ src/redux/slices/user-slice.ts
@@ -8,7 +8,7 @@ import {
   type EvaluationAdministration,
   type EvaluationAdministrationFilters,
 } from "../../types/evaluation-administration-type"
-import { axiosInstance } from "../../utils/axiosInstance"
+import { axiosInstance } from "../../utils/axios-instance"
 import { Loading } from "../../types/loadingType"
 
 export const getUserEvaluations = createAsyncThunk(
