+++ src/redux/store.ts
@@ -5,10 +5,10 @@ import authReducer from "./slices/auth-slice"
 import emailTemplateReducer from "./slices/email-template-slice"
 import evaluationAdministrationReducer from "./slices/evaluation-administration-slice"
 import evaluationAdministrationsReducer from "./slices/evaluation-administrations-slice"
-import evaluationResultReducer from "./slices/evaluationResultSlice"
-import evaluationResultsReducer from "./slices/evaluationResultsSlice"
+import evaluationResultReducer from "./slices/evaluation-result-slice"
+import evaluationResultsReducer from "./slices/evaluation-results-slice"
 import evaluationsReducer from "./slices/evaluations-slice"
-import evaluationTemplatesReducer from "./slices/evaluationTemplatesSlice"
+import evaluationTemplatesReducer from "./slices/evaluation-templates-slice"
 import evaluationTemplateContentsReducer from "./slices/evaluation-template-contents-slice"
 import externalUserReducer from "./slices/external-user-slice"
 import externalUsersReducer from "./slices/external-users-slice"
